big integer

+++ database/migrations/2016_07_29_171118_create_chatter_categories_table.php
@@ -9,7 +9,7 @@ class CreateChatterCategoriesTable extends Migration
     {
         Schema::create('chatter_categories', function (Blueprint $table) {
             $table->id('id');
-            $table->integer('parent_id')->unsigned()->nullable();
+            $table->bigInteger('parent_id')->unsigned()->nullable();
             $table->integer('order')->default(1);
             $table->string('name');
             $table->string('color', 20);
